Add files via upload
This Java code defines a class named "Oopss" which contains a method named "replace" that takes an integer array, an index, and a new element as input. The method replaces the element at the specified index in the array with the new element.

In the main method, an object of the Oopss class is created and initialized. An integer array is also defined and initialized with three values: 1, 2, and 3. The "replace" method of the Oopss object is then called, passing in the integer array, index "2" and the new element "78". The "replace" method updates the array by replacing the element at index "2" with "78".

Finally, the updated integer array is printed to the console using the Arrays.toString() method, which converts the array into a string representation. The output will be:

[1, 2, 78]